update plotting landmarks to remove VNC twist
* @schlegelp writes:
aschampion pointed out that the VNC was still slightly rotated counterclockwise. When I generated the first transform, I only tilted the VNC down but didn't correct for any other rotations.
I'm attaching a new set of landmarks that includes a rotation along the z-axis that will right both the VNC and the brain. The screenshot show the original in red and this new transform in green.

* see also https://flyconnectome.slack.com/archives/C02F6UCCU6B/p1754387741120049?thread_ts=1747645164.139839&cid=C02F6UCCU6B